fix(db): change notification.id to BIGINT for JPA Long mapping AB#1930

+++ backend/src/main/resources/db/migration/V003__create_notification.sql
@@ -6,7 +6,7 @@
 
 -- Notification transactional table for email records
 CREATE TABLE notification (
-    id INT IDENTITY(1,1) PRIMARY KEY,
+    id BIGINT IDENTITY(1,1) PRIMARY KEY,
     program_id INT NOT NULL,
     notification_type NVARCHAR(50) NOT NULL,
     recipient_email NVARCHAR(200) NOT NULL,
